Moved event processing into it's own class for better readability
